Move helper classes to fixture namespace
